update feature manage TypeOfBook, export CouponList to pdf v1.0
Description add, delete TypeOfBook, design modify addNewOrderScreen, binding username v1.0, add converter for currency

+++ Project1_BookStore/Utils/NumberToCurrencyConverter.cs
@@ -12,7 +12,7 @@ namespace Project1_BookStore.Utils
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            int number = (int)value;
+            decimal number = (decimal)value;
             var info = CultureInfo.GetCultureInfo("vi-VN");
             var result = String.Format(info, "{0:c}", number);
             return result;
